requirements.txt
accelerate==1.3.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
attrs==25.1.0
beautifulsoup4==4.13.4
bidict==0.23.1
blinker==1.9.0
branca==0.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
contourpy==1.3.1
cycler==0.12.1
datasets==3.5.0
decorator==5.1.1
decord==0.6.0
deepspeed==0.16.6
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
easydict==1.13
einops==0.8.1
executing==2.2.0
Farama-Notifications==0.0.4
filelock==3.17.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SocketIO==5.5.1
folium==0.19.4
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
gdown==5.2.0
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.44
google_search_results==2.4.2
gradio_client==1.7.0
greenlet==3.1.1
# Editable install with no version control (gym-cards-and-virl==0.0.2)
-e /path/RL_Heals_SFT
gymnasium==1.0.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
ipdb==0.13.13
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.18
langchain-core==0.3.34
langchain-text-splitters==0.3.6
langsmith==0.3.8
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
ninja==1.11.1.3
numpy @ file:///croot/numpy_and_numpy_base_1738078506782/work/dist/numpy-2.2.2-cp313-cp313-linux_x86_64.whl#sha256=7931599a0694936575a9e0bac58d750d6e5e8072bde698fbc948ce3073802feb
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.61.1
opencv-python==4.10.0
opencv-python-headless==4.10.0
orjson==3.10.15
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
peft==0.14.0
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
polyline==2.0.2
prettytable==3.14.0
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.1
pydantic==2.10.6
pydantic_core==2.27.2
pydevd==3.3.0
Pygments==2.19.1
pyparsing==3.2.1
PySocks==1.7.1
python-box==7.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.2
scikit-learn==1.7.1
scipy==1.16.0
seaborn==0.13.2
selenium==4.9.1
sentry-sdk==2.21.0
setproctitle==1.3.4
setuptools==72.1.0
shapely==2.0.7
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.38
stack-data==0.6.3
sympy==1.13.1
tenacity==9.0.0
termcolor==2.5.0
threadpoolctl==3.6.0
tokenizers==0.20.3
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
trio==0.28.0
trio-websocket==0.11.1
triton==3.2.0
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
urllib3==2.3.0
wandb==0.19.6
wcwidth==0.2.13
webdriver-manager==4.0.2
websockets==14.2
Werkzeug==3.1.3
wheel==0.44.0
wsproto==1.2.0
xlsxwriter==3.2.5
xxhash==3.5.0
xyzservices==2025.1.0
yarl==1.18.3
zstandard==0.23.0